apply new color version
